Split play_path.py into 3 files.
