+++ doc/main-page.hh
@@ -3,10 +3,10 @@
   Supervision of data transmission between HPP and SoT.
   It implements a finite state machine using Smach. There are three states:
   \li agimus.path_execution.play_path.WaitForInput
-  \li agimus.path_execution.play_path.InitializePath
-  \li agimus.path_execution.play_path.PlayPath
+  \li agimus.path_execution.play_path.InitializePath (Init)
+  \li agimus.path_execution.play_path.PlayPath (Play)
 
-  See packages agimus_sot, \ref agimus_hpp, agimus_vision and agimus_sot_msgs.
+  \image html figures/smach-viewer-inkscape.svg "Finite state machine."
 
   \todo Split file \c src/agimus/path_execution/play_path.py into 3 files with one class per file.
